feat: 매칭 결과에 따른 메시지를 생성하는 메서드를 구현한다.
